Implemented UI to add GitLab pipeline (#27).
As an experiment I tried to implement the UI completely with an AI-assistant, using prompts only. A few times, when the AI-assistant got "stuck", I looked at the code a bit so that I could give hints to the AI via the prompts, but I didn't write any code. I stopped when the happy path worked. This commit is the result of this "coding" session. For the record, I used Windsurf with Claude 3.7 Sonnet spending 38 prompt credits. At this time, for the model: 1 credit == 1 prompt.

Was it faster than writing the code directly? I'm not sure, especially when considering that I still have to review the code in the next session. Was it more fun? I don't think so. And I say this as someone who isn't particularly fond of SwiftUI.